- Further hardening and optimization

+++ GenOnlineService/Controllers/Lobby/LobbyController.cs
@@ -331,7 +331,13 @@ namespace GenOnlineService.Controllers
 								}
 
 								// register with daily stats
-								DailyStatsManager.RegisterOutcome(army, won);
+								// NOTE: only once per match - the outcome endpoint can be called repeatedly for the
+								// same match_id, and these counters are unconditional increments
+								bool bFirstOutcomeReportForMatch = sourceData.TryMarkMatchOutcomeReported(match_id);
+								if (bFirstOutcomeReportForMatch)
+								{
+									DailyStatsManager.RegisterOutcome(army, won);
+								}
 
 								// give them back signed URLs they need
 								result.screenshot_url = await S3CredentialManager.GetPresignedURL(EMetadataFileType.FILE_TYPE_SCREENSHOT, EScreenshotType.SCREENSHOT_TYPE_SCORESCREEN, match_id, user_id, slotIndexInLobby, LobbyCreationTime) ?? String.Empty;
@@ -550,22 +556,27 @@ namespace GenOnlineService.Controllers
 										// TODO: we should communicate the kick to the user...
 										Int64 KickedUserID = data["userid"].GetInt64();
 
-										_lobbyManager.LeaveSpecificLobby(KickedUserID, lobbyID);
-
-										// cleanup TURN credentials
-										TURNCredentialManager.DeleteCredentialsForUser(KickedUserID);
-
-										// clear our lobby ID
-										UserSession? sourceData = WebSocketManager.GetSessionFromUser(KickedUserID, EUserSessionType.GameClient); // user being kicked must be a game client
-
-										if (sourceData != null)
+										// the target must actually be in THIS lobby, otherwise a host could wipe the
+										// TURN credentials / lobby state of any arbitrary online player
+										if (lobby.GetMemberFromUserID(KickedUserID) != null)
 										{
-											sourceData.UpdateSessionLobbyID(-1);
-											// NOTE: We dont update the match history match ID here, that is done by the match history service
-										}
+											_lobbyManager.LeaveSpecificLobby(KickedUserID, lobbyID);
 
-										// we have to manually send to the kicked user... they won't get the dirty lobby update anymore
-										await lobby.DirtyRetransmitToSingleMember(KickedUserID);
+											// cleanup TURN credentials
+											TURNCredentialManager.DeleteCredentialsForUser(KickedUserID);
+
+											// clear our lobby ID
+											UserSession? sourceData = WebSocketManager.GetSessionFromUser(KickedUserID, EUserSessionType.GameClient); // user being kicked must be a game client
+
+											if (sourceData != null)
+											{
+												sourceData.UpdateSessionLobbyID(-1);
+												// NOTE: We dont update the match history match ID here, that is done by the match history service
+											}
+
+											// we have to manually send to the kicked user... they won't get the dirty lobby update anymore
+											await lobby.DirtyRetransmitToSingleMember(KickedUserID);
+										}
 									}
 								}
 								else if (field == ELobbyUpdateField.HOST_ACTION_SET_SLOT_STATE)
